Adds ABAB check bean refactoring
